Put the common methods in trait
